# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Ges-Comptes API",
        "description": "API pour la gestion des comptes bancaires",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/v1/clients/{identifier}": {
            "get": {
                "tags": [
                    "Clients"
                ],
                "summary": "Rechercher un client par téléphone ou NCI",
                "description": "Rechercher un client par téléphone ou NCI",
                "operationId": "6b4e6ccf06dbc08e7f1d5ed8e332091d",
                "parameters": [
                    {
                        "name": "identifier",
                        "in": "path",
                        "description": "Numéro de téléphone ou numéro de carte d'identité (NCI)",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Client trouvé",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Client trouvé"
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/ClientResource"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Client non trouvé",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "error": {
                                            "properties": {
                                                "code": {
                                                    "type": "string",
                                                    "example": "CLIENT_NOT_FOUND"
                                                },
                                                "message": {
                                                    "type": "string",
                                                    "example": "Aucun client trouvé avec ce numéro de téléphone ou NCI"
                                                },
                                                "details": {
                                                    "type": "object"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Erreur serveur",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Erreur serveur"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/clients/{client}/comptes": {
            "get": {
                "tags": [
                    "Clients"
                ],
                "summary": "Lister les comptes d'un client",
                "description": "Lister les comptes d'un client",
                "operationId": "06d10c4ece779261692c95944eb8c4a2",
                "parameters": [
                    {
                        "name": "client",
                        "in": "path",
                        "description": "ID du client",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Comptes du client récupérés",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string"
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/CompteResource"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Client non trouvé",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "error": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "ClientResource": {
                "properties": {
                    "id": {
                        "type": "string",
                        "example": "uuid"
                    },
                    "nom": {
                        "type": "string",
                        "example": "Doe"
                    },
                    "prenom": {
                        "type": "string",
                        "example": "John"
                    },
                    "email": {
                        "type": "string",
                        "format": "email"
                    },
                    "telephone": {
                        "type": "string",
                        "example": "+221771234567"
                    },
                    "cni": {
                        "type": "string",
                        "example": "123456789012"
                    },
                    "adresse": {
                        "type": "string"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
            "CompteResource": {
                "properties": {
                    "id": {
                        "type": "string",
                        "example": "uuid"
                    },
                    "numero_compte": {
                        "type": "string",
                        "example": "CPT-ABC123"
                    },
                    "titulaire": {
                        "type": "string",
                        "example": "John Doe"
                    },
                    "type": {
                        "type": "string",
                        "enum": [
                            "epargne",
                            "cheque"
                        ]
                    },
                    "solde_initial": {
                        "type": "number",
                        "format": "float",
                        "example": 10000
                    },
                    "solde": {
                        "type": "number",
                        "format": "float",
                        "example": 1000.5
                    },
                    "devise": {
                        "type": "string",
                        "example": "XOF"
                    },
                    "date_creation": {
                        "type": "string",
                        "format": "date",
                        "example": "2023-01-01"
                    },
                    "statut": {
                        "type": "string",
                        "enum": [
                            "actif",
                            "bloque",
                            "ferme"
                        ]
                    },
                    "client_id": {
                        "type": "string",
                        "example": "uuid"
                    },
                    "informations_blocage": {
                        "description": "Informations de blocage affichées pour les comptes épargne (même si non bloqués actuellement)",
                        "properties": {
                            "motifBlocage": {
                                "type": "string",
                                "example": "Blocage pour vérification"
                            },
                            "dateBlocage": {
                                "type": "string",
                                "format": "date-time"
                            },
                            "dateDeblocagePrevue": {
                                "type": "string",
                                "format": "date-time"
                            },
                            "motifDeblocage": {
                                "type": "string",
                                "nullable": true
                            },
                            "dateDeblocage": {
                                "type": "string",
                                "format": "date-time",
                                "nullable": true
                            }
                        },
                        "type": "object"
                    },
                    "metadata": {
                        "type": "object"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
...
        }
    },
...
}